feat: createSudokuSettings
